Revert "refactor : 쿠키 처리 파일 분리(utils/cookieUtils.ts)"
This reverts commit efce49a3d0dc10a8d074af9555c9528c9ef7dfaf.
